app/requirements.txt
numpy~=1.19.5
pandas~=1.1.5
streamlit~=0.76.0
annoy==1.17.0
h5py==2.10.0
matplotlib==3.2.2
seaborn==0.11.0
Pillow==8.1.0